+++ app/(tabs)/mypage.tsx
@@ -1,0 +1,27 @@
+import { useRouter } from 'expo-router';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function MyPage() {
+  const router = useRouter();
+
+  return (
+    <>
+      <View style={styles.container}>
+        <Text style={styles.text}>마이 페이지</Text>
+      </View>
+    </>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#fff',
+  },
+  text: {
+    fontSize: 20,
+    fontWeight: 'bold',
+  },
+});
