+++ app/(tabs)/setting.tsx
@@ -1,28 +1,3 @@
-import { useRouter } from 'expo-router';
-import { StyleSheet, Text, View } from 'react-native';
-import BottomTabBar from '../../components/common/BottomTabBar';
+import Settingpage from '@/app/(tabs)/myPage/myPage';
 
-export default function SettingPage() {
-  const router = useRouter();
-
-  return (
-    <>
-      <View style={styles.container}>
-        <Text style={styles.text}>세팅 페이지</Text>
-      </View>
-    </>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#fff',
-  },
-  text: {
-    fontSize: 20,
-    fontWeight: 'bold',
-  },
-});
+export default Settingpage;
